refactor: vite.config.ts, minor changes in tests
