+++ src/components/Header/Header.test.tsx
@@ -1,6 +1,6 @@
 import { describe, test, expect } from 'vitest';
 import { render, screen } from '@testing-library/react';
-import Header from './Header';
+import Header from '.';
 import React from 'react';
 import { MemoryRouter } from 'react-router-dom';
