+++ src/components/SearchBar/SearchBar.test.tsx
@@ -2,7 +2,7 @@ import { describe, test, expect } from 'vitest';
 import { render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import '@testing-library/jest-dom';
-import SearchBar from './SearchBar';
+import SearchBar from '.';
 import React from 'react';
 
 describe('<SearchBar />', () => {
